requirements.txt
watchdog>=6.0
numpy
Pillow
